+++ R/xgvis.R
@@ -1,31 +1,33 @@
 xgvis <-
-function(dmat       = NULL,
-         edges      = NULL,
-         pos        = NULL,
-         rowlab     = dimnames(dmat)[[1]],
-         colors     = NULL,
-         glyphs     = NULL,
-         erase      = NULL,
-         lines      = NULL,
-         linecolors = NULL,
-         resources  = NULL,
-         display    = NULL)
+function(dmat	= NULL,
+	 edges	= NULL,
+	 pos	= NULL,
+	 rowlab = dimnames(dmat)[[1]],
+	 colors = NULL,
+	 glyphs = NULL,
+	 erase	= NULL,
+	 lines	= NULL,
+	 linecolors = NULL,
+	 resources  = NULL,
+	 display    = NULL,
+	 keep	= FALSE,
+	 fprefix= "xgvis-")
 {
   if (is.null(edges) && is.null(pos) && is.null(dmat))
     stop("One of dmat, edges, or pos must be present")
 
-  basefile <- tempfile("xgvis")
+  basefile <- tempfile(fprefix)
 
   ## distance matrix ###
   if (!is.null(dmat)) {
     dmat <- eval(dmat)
     if (any(isinf <- is.infinite(dmat[!is.na(dmat)]))) {
-        warning("xgvis can't handle Inf's in dmat; replaced with NA")
-        dmat[isinf] <- NA
+	warning("xgvis can't handle Inf's in dmat; replaced with NA")
+	dmat[isinf] <- NA
     }
     dfile <- paste(basefile, ".dist", sep="")
     write(t(dmat), file = dfile, ncolumns = ncol(dmat))
-    on.exit(unlink(dfile), add=TRUE)
+    if(!keep) on.exit(unlink(dfile), add=TRUE)
   }
 
   ## Edges ###
@@ -37,29 +39,29 @@ function(dmat       = NULL,
     if (nrow(edges) > 0) {
 	write(t(edges), file = edgesfile, ncol=2)
     }
-    on.exit(unlink(edgesfile), add=TRUE)
+    if(!keep) on.exit(unlink(edgesfile), add=TRUE)
   }
 
   ## position matrix ###
   if (!is.null(pos)) {
     pos <- eval(pos)
     if (any(isinf <- is.infinite(pos[!is.na(pos)]))) {
-        warning("xgvis can't handle Inf's in pos; replaced with NA")
-        pos[isinf] <- NA
+	warning("xgvis can't handle Inf's in pos; replaced with NA")
+	pos[isinf] <- NA
     }
     pfile <- paste(basefile, ".pos", sep="")
     write(t(pos), file = pfile, ncolumns = ncol(pos))
-    on.exit(unlink(pfile), add = TRUE)
+    if(!keep) on.exit(unlink(pfile), add = TRUE)
   }
 
   ## Row / Case labels ###
   if (!is.null(rowlab)) {
       if (!is.vector(rowlab) || !is.character(rowlab))# check data type
-          stop("The `rowlab' argument needs to be a character vector")
+	  stop("The `rowlab' argument needs to be a character vector")
       if (!missing(rowlab) && length(rowlab) != NROW(dmat))
-          stop("`rowlab' has wrong length (not matching NROW(dmat))")
+	  stop("`rowlab' has wrong length (not matching NROW(dmat))")
       cat(rowlab, file = (rowfile <- paste(basefile, ".row", sep="")), sep="\n")
-      on.exit(unlink(rowfile), add = TRUE)
+      if(!keep) on.exit(unlink(rowfile), add = TRUE)
   }
 
   ## Colors ###
@@ -70,7 +72,7 @@ function(dmat       = NULL,
 
     colorfile <- paste(basefile, ".colors", sep="")
     write(colors, file = colorfile, ncol=1)
-    on.exit(unlink(colorfile), add = TRUE)
+    if(!keep) on.exit(unlink(colorfile), add = TRUE)
   }
 
   ## Glyphs ###
@@ -81,7 +83,7 @@ function(dmat       = NULL,
 
     glyphfile <- paste(basefile, ".glyphs", sep="")
     write(glyphs, file = glyphfile, ncol=1)
-    on.exit(unlink(glyphfile), add = TRUE)
+    if(!keep) on.exit(unlink(glyphfile), add = TRUE)
   }
 
   ## Erase ###
@@ -92,7 +94,7 @@ function(dmat       = NULL,
 
     erasefile <- paste(basefile, ".erase", sep="")
     write(erase, file = erasefile, ncol=1)
-    on.exit(unlink(erasefile), add = TRUE)
+    if(!keep) on.exit(unlink(erasefile), add = TRUE)
   }
 
   ## Connected lines ###
@@ -104,7 +106,7 @@ function(dmat       = NULL,
     linesfile <- paste(basefile, ".lines", sep="")
     if (nrow(lines) > 0) {
       write(t(lines), file = linesfile, ncol=2)
-      on.exit(unlink(linesfile), add = TRUE)
+      if(!keep) on.exit(unlink(linesfile), add = TRUE)
     }
   }
 
@@ -116,7 +118,7 @@ function(dmat       = NULL,
 
     linecolorfile <- paste(basefile, ".linecolors", sep="")
     write(linecolors, file = linecolorfile, ncol=1)
-    on.exit(unlink(linecolorfile), add = TRUE)
+    if(!keep) on.exit(unlink(linecolorfile), add = TRUE)
   }
 
   ## Resources ###
@@ -127,11 +129,8 @@ function(dmat       = NULL,
 
     resourcefile <- paste(basefile, ".resources", sep="")
     write(resources, file = resourcefile, ncol=1)
-    on.exit(unlink(resourcefile), add = TRUE)
+    if(!keep) on.exit(unlink(resourcefile), add = TRUE)
   }
-
-## DEBUGGING: Keep all the tempfiles by
-##> on.exit()
 
 ### Note to installer:
 ### Here you need to specify the path to the xgvis executable / batch file
@@ -143,8 +142,8 @@ function(dmat       = NULL,
   ## invisible(dos(command, multi= F, minimized=T, output.to.S=F, translate=T))
   s <- system(command, FALSE)
 
-  ## Now wait a bit before unlinking all the files via  on.exit(.) :
-  system("sleep 3")
+  ## Now wait a bit before unlinking all the files via on.exit(.) :
+  if(!keep) system("sleep 3")
   invisible(s)
 }
